completed react-redux
