+++ client/components/main.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {withRouter, Link} from 'react-router-dom'
 import {logout} from '../store'
-import Sidebar from './Sidebar'
-
 import Navbar from './Navbar'
+import AllProducts from './AllProducts'
+import Sidebar from './Sidebar'
 
 /**
  * COMPONENT
@@ -18,29 +18,9 @@ const Main = (props) => {
 
   return (
     <div>
-
+      <Navbar />
       <Sidebar />
-    // <div>
-    //   <h1>BOILERMAKER</h1>
-    //   <nav>
-    //     {
-    //       isLoggedIn
-    //         ? <div>
-    //           {/* The navbar will show these links after you log in */}
-    //           <Link to='/home'>Home</Link>
-    //           <a href='#' onClick={handleClick}>Logout</a>
-    //         </div>
-    //         : <div>
-    //           {/* The navbar will show these links before you log in */}
-    //           <Link to='/login'>Login</Link>
-    //           <Link to='/signup'>Sign Up</Link>
-    //         </div>
-    //     }
-    //   </nav>
-    //   <hr />
-    //   {children}
-    // </div>
-    <Navbar />
+      <AllProducts />
     </div>
   )
 }
@@ -77,7 +57,7 @@ Main.propTypes = {
 
 
       // <h1>BOILERMAKER</h1>
-      
+
       // <nav>
       //   {
       //     isLoggedIn
