+++ client/index.js
@@ -4,7 +4,6 @@ import ReactDOM from 'react-dom'
 import {Provider} from 'react-redux'
 import store from './store'
 import Routes from './routes'
-import AllProducts from './components/AllProducts/main.js'
 
 // establishes socket connection
 import './socket'
@@ -14,7 +13,7 @@ import './index.scss'
 
 ReactDOM.render(
   <Provider store={store}>
-    <AllProducts />
+    <Routes />
   </Provider>,
   document.getElementById('app')
 )
